feat: add DocSearch v3
