+++ docusaurus.config.js
@@ -20,7 +20,9 @@ module.exports = {
     algolia: {
       apiKey: '25626fae796133dc1e734c6bcaaeac3c',
       indexName: 'docsearch',
-      algoliaOptions: {}, // Optional, if provided by Algolia
+      searchParameters: {
+        hitsPerPage: 5,
+      },
     },
     navbar: {
       logo: {
