+++ package.json
@@ -15,9 +15,7 @@
     "@algolia/ui-library": "^1.0.0-beta.109",
     "@docusaurus/core": "2.0.0-alpha.48",
     "@docusaurus/preset-classic": "2.0.0-alpha.48",
-    "@francoischalifour/autocomplete-core": "^1.0.0-alpha.10",
-    "@francoischalifour/autocomplete-preset-algolia": "^1.0.0-alpha.10",
-    "@francoischalifour/autocomplete.js": "^1.0.0-alpha.9",
+    "@francoischalifour/docsearch-react": "1.0.0-alpha.13",
     "algoliasearch": "^4.1.0",
     "classnames": "^2.2.6",
     "docsearch.js": "^2.6.3",
