--- src/theme/DocSearch-react/esm/Hit.js
@@ -1,8 +1,0 @@
-import React from 'react';
-export function Hit(_ref) {
-  var hit = _ref.hit,
-      children = _ref.children;
-  return React.createElement("a", {
-    href: hit.url
-  }, children);
-}
