--- src/theme/DocSearch-react/esm/Results.js
@@ -1,119 +1,0 @@
-function _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _nonIterableRest(); }
-
-function _nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance"); }
-
-function _iterableToArrayLimit(arr, i) { if (!(Symbol.iterator in Object(arr) || Object.prototype.toString.call(arr) === "[object Arguments]")) { return; } var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"] != null) _i["return"](); } finally { if (_d) throw _e; } } return _arr; }
-
-function _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }
-
-function _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }
-
-import React from 'react';
-import { Snippet } from './Snippet';
-export function Results(props) {
-  if (!props.suggestion || props.suggestion.items.length === 0) {
-    return null;
-  }
-
-  return React.createElement("section", {
-    className: "DocSearch-Hits"
-  }, React.createElement("div", {
-    className: "DocSearch-Hit-source"
-  }, props.title), React.createElement("ul", props.getMenuProps(), props.suggestion.items.map(function (item, index) {
-    return React.createElement(Result, _extends({
-      key: [props.title, item.objectID].join(':'),
-      item: item,
-      index: index
-    }, props));
-  })));
-}
-
-function Result(_ref) {
-  var item = _ref.item,
-      index = _ref.index,
-      renderIcon = _ref.renderIcon,
-      renderAction = _ref.renderAction,
-      getItemProps = _ref.getItemProps,
-      onItemClick = _ref.onItemClick,
-      suggestion = _ref.suggestion,
-      hitComponent = _ref.hitComponent;
-
-  var _React$useState = React.useState(false),
-      _React$useState2 = _slicedToArray(_React$useState, 2),
-      isDeleting = _React$useState2[0],
-      setIsDeleting = _React$useState2[1];
-
-  var _React$useState3 = React.useState(false),
-      _React$useState4 = _slicedToArray(_React$useState3, 2),
-      isFavoriting = _React$useState4[0],
-      setIsFavoriting = _React$useState4[1];
-
-  var action = React.useRef(null);
-  var Hit = hitComponent;
-
-  function runDeleteTransition(cb) {
-    setIsDeleting(true);
-    action.current = cb;
-  }
-
-  function runFavoriteTransition(cb) {
-    setIsFavoriting(true);
-    action.current = cb;
-  }
-
-  return React.createElement("li", _extends({
-    className: ['DocSearch-Hit', item.__docsearch_parent && 'DocSearch-Hit--Child', isDeleting && 'DocSearch-Hit--deleting', isFavoriting && 'DocSearch-Hit--favoriting'].filter(Boolean).join(' '),
-    onTransitionEnd: function onTransitionEnd() {
-      if (action.current) {
-        action.current();
-      }
-    }
-  }, getItemProps({
-    item: item,
-    source: suggestion.source,
-    onClick: function onClick() {
-      onItemClick(item);
-    }
-  })), React.createElement(Hit, {
-    hit: item
-  }, React.createElement("div", {
-    className: "DocSearch-Hit-Container"
-  }, renderIcon({
-    item: item,
-    index: index
-  }), item.hierarchy[item.type] && item.type === 'lvl1' && React.createElement("div", {
-    className: "DocSearch-Hit-content-wrapper"
-  }, React.createElement(Snippet, {
-    className: "DocSearch-Hit-title",
-    hit: item,
-    attribute: "hierarchy.lvl1"
-  }), item.content && React.createElement(Snippet, {
-    className: "DocSearch-Hit-path",
-    hit: item,
-    attribute: "content"
-  })), item.hierarchy[item.type] && (item.type === 'lvl2' || item.type === 'lvl3' || item.type === 'lvl4' || item.type === 'lvl5' || item.type === 'lvl6') && React.createElement("div", {
-    className: "DocSearch-Hit-content-wrapper"
-  }, React.createElement(Snippet, {
-    className: "DocSearch-Hit-title",
-    hit: item,
-    attribute: "hierarchy.".concat(item.type)
-  }), React.createElement(Snippet, {
-    className: "DocSearch-Hit-path",
-    hit: item,
-    attribute: "hierarchy.lvl1"
-  })), item.type === 'content' && React.createElement("div", {
-    className: "DocSearch-Hit-content-wrapper"
-  }, React.createElement(Snippet, {
-    className: "DocSearch-Hit-title",
-    hit: item,
-    attribute: "content"
-  }), React.createElement(Snippet, {
-    className: "DocSearch-Hit-path",
-    hit: item,
-    attribute: "hierarchy.lvl1"
-  })), renderAction({
-    item: item,
-    runDeleteTransition: runDeleteTransition,
-    runFavoriteTransition: runFavoriteTransition
-  }))));
-}
