--- src/theme/DocSearch-react/esm/icons/GoToExternalIcon.js
@@ -1,11 +1,0 @@
-import React from 'react';
-export function GoToExternal() {
-  return React.createElement("svg", {
-    width: "20",
-    height: "20"
-  }, React.createElement("path", {
-    d: "M5 6v9h9v-3a1 1 0 112 0v4l-1 1H4l-1-1V5l1-1h4a1 1 0 110 2H5zm5 5a1 1 0 11-1-1l5-6h-3a1 1 0 110-2h6a1 1 0 011 1v6a1 1 0 11-2 0V6l-6 5z",
-    fill: "currentColor",
-    fillRule: "nonzero"
-  }));
-}
