--- src/theme/DocSearch-react/esm/icons/LoadingIcon.js
@@ -1,28 +1,0 @@
-import React from 'react';
-export function LoadingIcon() {
-  return React.createElement("svg", {
-    viewBox: "0 0 38 38",
-    stroke: "currentColor",
-    strokeOpacity: ".5"
-  }, React.createElement("g", {
-    fill: "none",
-    fillRule: "evenodd"
-  }, React.createElement("g", {
-    transform: "translate(1 1)",
-    strokeWidth: "2"
-  }, React.createElement("circle", {
-    strokeOpacity: ".3",
-    cx: "18",
-    cy: "18",
-    r: "18"
-  }), React.createElement("path", {
-    d: "M36 18c0-9.94-8.06-18-18-18"
-  }, React.createElement("animateTransform", {
-    attributeName: "transform",
-    type: "rotate",
-    from: "0 18 18",
-    to: "360 18 18",
-    dur: "1s",
-    repeatCount: "indefinite"
-  })))));
-}
