--- src/theme/DocSearch-react/esm/icons/NoResultsIcon.js
@@ -1,15 +1,0 @@
-import React from 'react';
-export function NoResultsIcon() {
-  return React.createElement("svg", {
-    width: "40",
-    height: "40",
-    viewBox: "0 0 20 20",
-    fill: "none",
-    fillRule: "evenodd",
-    stroke: "currentColor",
-    strokeLinecap: "round",
-    strokeLinejoin: "round"
-  }, React.createElement("path", {
-    d: "M15.5 4.8c2 3 1.7 7-1 9.7h0l4.3 4.3-4.3-4.3a7.8 7.8 0 01-9.8 1m-2.2-2.2A7.8 7.8 0 0113.2 2.4M2 18L18 2"
-  }));
-}
