--- src/theme/DocSearch-react/esm/icons/ResetIcon.js
@@ -1,15 +1,0 @@
-import React from 'react';
-export function ResetIcon() {
-  return React.createElement("svg", {
-    width: "20",
-    height: "20",
-    viewBox: "0 0 20 20"
-  }, React.createElement("path", {
-    d: "M10 10l5.09-5.09L10 10l5.09 5.09L10 10zm0 0L4.91 4.91 10 10l-5.09 5.09L10 10z",
-    stroke: "currentColor",
-    fill: "none",
-    fillRule: "evenodd",
-    strokeLinecap: "round",
-    strokeLinejoin: "round"
-  }));
-}
