--- src/theme/DocSearch-react/esm/icons/SelectIcon.js
@@ -1,18 +1,0 @@
-import React from 'react';
-export function SelectIcon() {
-  return React.createElement("svg", {
-    width: "20",
-    height: "20",
-    viewBox: "0 0 20 20"
-  }, React.createElement("g", {
-    stroke: "currentColor",
-    fill: "none",
-    fillRule: "evenodd",
-    strokeLinecap: "round",
-    strokeLinejoin: "round"
-  }, React.createElement("path", {
-    d: "M18 3v4c0 2-2 4-4 4H2"
-  }), React.createElement("path", {
-    d: "M8 17l-6-6 6-6"
-  })));
-}
