--- src/theme/DocSearch-react/esm/icons/index.js
@@ -1,10 +1,0 @@
-export * from './GoToExternalIcon';
-export * from './LoadingIcon';
-export * from './RecentIcon';
-export * from './ResetIcon';
-export * from './SearchIcon';
-export * from './SelectIcon';
-export * from './SourceIcon';
-export * from './StarIcon';
-export * from './ErrorIcon';
-export * from './NoResultsIcon';
