--- src/theme/DocSearch-react/esm/index.js
@@ -1,2 +1,0 @@
-export * from './DocSearch';
-export * from './SearchButton';
