--- src/theme/DocSearch-react/esm/types/index.js
@@ -1,3 +1,0 @@
-export * from './DocSearchHit';
-export * from './InternalDocSearchHit';
-export * from './StoredDocSearchHit';
