--- src/theme/DocSearch-react/esm/utils/index.js
@@ -1,2 +1,0 @@
-export * from './groupBy';
-export * from './noop';
