--- src/theme/DocSearch-react/esm/utils/noop.js
@@ -1,1 +1,0 @@
-export function noop() {}
