--- src/theme/DocSearch-react/esm/version.js
@@ -1,1 +1,0 @@
-export var version = '1.0.0-alpha.10';
