--- src/theme/SearchBarr/sass/styles.sass
@@ -1,248 +1,0 @@
-
-@at-root :root
-  --algolia-autocomplete-input-color: var(--ifm-link-color)
-  --algolia-autocomplete-highlight-color: var(--ifm-link-color)
-  --algolia-autocomplete-background-highlight-color: var(--ifm-color-emphasis-200)
-  --algolia-autocomplete-border-color: rgba(197, 201, 224, 0.4)
-  --algolia-docsearch-lvl0-color: var(--ifm-link-color)
-  --algolia-docsearch-logo-color-outer: #5468ff
-  --algolia-docsearch-logo-color-inner: #fff
-  --algolia-docsearch-text-color: var(--ifm-color-emphasis-600)
-  --algolia-docsearch-tree-color: rgba(197, 201, 224, 0.8)
-  --algolia-autocomplete-shadow-color: rgba(197, 201, 224, 0.8)
-  --algolia-autocomplete-shadow-color-alpha: rgba(197, 201, 224, 0.4)
-  --algolia-autocomplete-searchbox-background-color: var(--ifm-color-emphasis-200)
-
-html[data-theme='dark']
-  --algolia-autocomplete-input-color: var(--ifm-link-color)
-  --algolia-autocomplete-background-highlight-color: var(--ifm-navbar-background-color)
-  --algolia-autocomplete-searchbox-background-color: var(--ifm-background-color)
-  --algolia-docsearch-tree-color: var(--solstice-blue)
-  --algolia-autocomplete-border-color: var(--solstice-blue)
-  --algolia-autocomplete-shadow-color: rgba(0, 0, 0, 0.8)
-  --algolia-autocomplete-shadow-color-alpha: rgba(0, 0, 0, 0.4)
-  --algolia-docsearch-logo-color-outer: var(--white)
-  --algolia-docsearch-logo-color-inner: var(--ifm-background-color)
-
-.algolia-autocomplete
-  &-form
-    display: flex
-    position: relative
-    margin: 0 0 0 1.4em
-    padding: 0 8px
-    border-radius: 6px
-    align-items: center
-    background: var(--algolia-autocomplete-searchbox-background-color)
-  &-input
-    appearance: none
-    width: 100%
-    height: 100%
-    padding: 0.7rem
-    border: none
-    outline: none
-    font: inherit
-    color: var(--ifm-color-emphasis-800)
-    background: inherit
-    &::-webkit-search-cancel-button,
-    &::-webkit-search-decoration,
-    &::-webkit-search-results-button,
-    &::-webkit-search-results-decoration
-      display: none
-  &-loadingIndicator
-    display: none
-  &-magnifierLabel,
-  &-reset
-    display: flex
-    height: 2rem
-    width: 2rem
-    color: var(--algolia-autocomplete-input-color)
-    align-items: center
-    justify-content: center
-  &--stalled .algolia-autocomplete-magnifierLabel
-    align-items: center
-    justify-content: center
-    display: none
-  &-loadingIndicator
-    svg
-      height: 18px
-      width: 18px
-  &-magnifierLabel
-    padding-top: 4px
-    svg
-      height: 18px
-      width: 18px
-  &-searchbox
-    display: flex
-    flex: 1
-  &-completion
-    appearance: none
-    position: absolute
-    width: 100%
-    height: 100%
-    padding: 0.7rem
-    border: none
-    outline: none
-    opacity: 0.4
-    font: inherit
-    pointer-events: none
-    background: inherit
-  &-suggestions
-    &-item
-      cursor: default
-      a
-        color: inherit
-        display: block
-      &[aria-selected='true']
-        background-color: var(--algolia-autocomplete-background-highlight-color)
-    &-footer > div
-      margin-top: 1.2em
-      border-top: solid 1px var(--algolia-autocomplete-border-color)
-      align-items: center
-      display: flex
-      flex-wrap: wrap
-      height: 3em
-      width: 100%
-      justify-content: space-between
-    mark
-      background: none
-      color: var(--algolia-autocomplete-highlight-color)
-  &-reset
-    appearance: none
-    background: none
-    border: none
-    border-radius: 50%
-    cursor: pointer
-    right: 0
-    position: absolute
-    svg
-      height: 10px
-      width: 10px
-    &[hidden]
-      display: none
-    &:focus
-      background: var(--algolia-autocomplete-background-light-color)
-      color: var(--algolia-autocomplete-highlight-color)
-      outline: none
-  &-dropdown
-    background-color: var(--ifm-background-color)
-    box-shadow: 0 1px 4px 0 var(--algolia-autocomplete-shadow-color), 0 3px 15px 0 var(--algolia-autocomplete-shadow-color-alpha)
-    margin-top: 8px
-    overflow: scroll
-    position: absolute
-    width: 100%
-    z-index: 1
-    &-container
-      position: relative
-      padding: 0 1.4rem 0
-    &::-webkit-scrollbar
-      display: none
-    ul
-      list-style: none
-      margin: 0
-      padding: 0
-    a
-      text-decoration: none
-    @media (min-width: 540px)
-      max-width: 500px
-      border-radius: 6px
-  &-commands
-    display: flex
-    color: var(--algolia-docsearch-text-color)
-    &-border
-      color: var(--algolia-autocomplete-border-color)
-    &-description
-      font-size: 0.75rem
-      letter-spacing: 0
-      margin-left: 0.25rem
-    li
-      display: flex
-      align-items: center
-      margin-right: 1rem
-    svg
-      height: 1rem
-      width: 1rem
-    svg + svg
-      margin-left: 0.125rem
-  &-logo
-    svg
-      position: relative
-      top: 4px
-      color: var(--algolia-docsearch-text-color)
-      height: 1.2rem
-    &-brand-outer
-      fill: var(--algolia-docsearch-logo-color-outer)
-    &-brand-inner
-      fill: var(--algolia-docsearch-logo-color-inner)
-
-.DSV3-cat-separator
-  width: 100%
-  font-size: 0.85em
-  font-weight: 500
-  margin-top: 1.8em
-  background: var(--ifm-background-color)
-  color: var(--algolia-docsearch-lvl0-color)
-  &::after
-    display: block
-    width: 100%
-    height: 0.4em
-    content: ''
-    border-left: solid 1px var(--algolia-docsearch-tree-color)
-
-.DSV3-title
-  font-size: 0.95em
-  display: block
-  overflow: hidden
-  width: 100%
-  white-space: nowrap
-  text-overflow: ellipsis
-  color: var(--ifm-color-emphasis-900)
-  &::before
-    display: inline-block
-    width: 0.6em
-    height: 0.9em
-    margin-right: 0.2em
-    vertical-align: top
-    content: ''
-
-.DSV3-text
-  margin-left: 1.2em
-  padding-bottom: 0.4em
-  line-height: 1.2em
-  font-size: 0.75em
-  color: var(--ifm-color-emphasis-600)
-
-.DSV3-lvl1
-  line-height: 1.6em
-  border-left: solid 1px var(--algolia-docsearch-tree-color)
-  .DSV3-title
-    &::before
-      border-bottom: solid 1px var(--algolia-docsearch-tree-color)
-  &.DSV3-lastLvl0
-    .DSV3-title
-      border-left: none
-      &::before
-        border-left: solid 1px var(--algolia-docsearch-tree-color)
-
-.DSV3-lvl2,
-.DSV3-lvl3,
-.DSV3-content
-  line-height: 1.8em
-  padding-left: 1em
-  border-left: solid 1px var(--algolia-docsearch-tree-color)
-  .DSV3-title
-    border-left: solid 1px var(--algolia-docsearch-tree-color)
-    &::before
-      border-bottom: solid 1px var(--algolia-docsearch-tree-color)
-
-.DSV3-lastLvl0
-  border: none
-  &.DSV3-lastLvl
-    .DSV3-title
-      border: none
-
-.DSV3-lastLvl
-  .DSV3-title
-    border: none
-    &::before
-      border-bottom: solid 1px var(--algolia-docsearch-tree-color)
-      border-left: solid 1px var(--algolia-docsearch-tree-color)
